+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog
 
-All Notable changes to `json-guard-cli` will be documented in this file.
+## 0.3.0 - 2017-04-30
 
-Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
+- Updated to use json-guard 1.0.0 and json-reference 1.0.0
+- Changed the validate headers from keyword, message, pointer, value to message, schema path, data path, cause.
 
 ## 0.2.1 - 2016-12-23